Provide methods to get a user's display name and search by display name.
The display name in the default implementation simply returns the regular
username. The get and search methods exist so that third party Bundles can
make the semantic distinction between a username and a display name without
having to know whether they are different at all. The username is to be
used for login and other internal use, while the display name should be
shown in UIs. So when a user searches by a user's name, it will be a search
by display name. If one then chooses to overwrite displayName with e.g.
firstname + lastname, third party Bundles continue to work correctly
without modification.