Renamed a parameter in UserCreator
I renamed a parameter in FOSUserBundle\UserCreator::create.
The affected param was renamed from $inactive to $active.

There is no reason to use a setEnabled() method on the user object,
which obviously expects a param $active, but on the other hand
call setEnabled(!$isInactive) method on the user object, which
translates to 'not not active'.

I really believe that we should unify this throughout the whole
bundle and therefore change this.