Update index.md
From what I have seen, it will NOT install the bundle by just updating. If you were to just run the update, composer says that it sees the new item listed,but since it does not already exist, it will ignore it. By ignoring it, it does not install it.