Update groups.rst
The extension of the Group from the FOSBundle must be from the Entity, not from the Model. If you extend the model, the required fields "name" and "roles" will not be created in the database. Took me an hour to find this error, I hope it can be changed soon.