this was just inverted.
user provider is supposed to support the class passed as parameter if its user manager class is the same or a subclass but the code is inverting the test. it only return true if if the manager is for a super class. see http://php.net/manual/es/function.is-subclass-of.php